1. Move com.taoswork.tallybook.dynamic.dataservice.core.dataio -> com.taoswork.tallybook.dynamic.dataio.reference 2. Rename EntityInstanceTranslator -> EntityTranslator 3. Rename EntityValueCopierManager -> EntityCopierManager 4. Rename CrossEntityManagerPersistableCopier -> EntityCopier 5. Rename EntityValueCopierService -> EntityCopierService 6. RawEntityInfoBuilder build group/tab info ahead of field info 7. Make ImmutableClassMetadata.referencingClassMetadata immutable 8. Prepare Collection-type support

+++ test/test-framework/src/main/java/com/taoswork/tallybook/testframework/domain/business/IDepartment.java
@@ -9,9 +9,9 @@ import java.util.List;
 import java.util.Map;
 
 public interface IDepartment extends Persistable {
-    int getId();
+    Long getId();
 
-    void setId(int id);
+    void setId(Long id);
 
     String getName();
